+++ compose/ui/ui/src/desktopTest/kotlin/androidx/compose/ui/awt/ComposePanelTest.kt
@@ -622,66 +622,6 @@ class ComposePanelTest {
     }
 
     @Test
-    fun focusableNonBlockingPopup_withComponentLayerType_passesClicksThrough() {
-        ComposeFeatureFlags.layerType.withOverride(LayerType.OnComponent) {
-            ComposeFeatureFlags.useSwingGraphicsInComposePanel.withOverride(true) {
-                val window = JFrame()
-                try {
-                    runApplicationTest {
-                        var showPopup by mutableStateOf(false)
-                        var backgroundClickCount = 0
-
-                        val composePanel = ComposePanel()
-                        composePanel.setContent {
-                            Box(
-                                Modifier
-                                    .size(200.dp)
-                                    .background(Color.Yellow)
-                                    .clickable { backgroundClickCount++ }
-                            )
-
-                            if (showPopup) {
-                                Popup(
-                                    properties = PopupProperties(
-                                        focusable = true,
-                                        dismissOnClickOutside = false,
-                                        consumePointerInputOutside = false,
-                                    ),
-                                ) {
-                                    Box(
-                                        Modifier
-                                            .size(50.dp)
-                                            .background(Color.Blue)
-                                    )
-                                }
-                            }
-                        }
-
-                        composePanel.windowContainer = window.layeredPane
-
-                        window.contentPane.add(composePanel, BorderLayout.CENTER)
-                        window.pack()
-                        window.isVisible = true
-
-                        awaitIdle()
-
-                        showPopup = true
-                        awaitIdle()
-
-                        window.layeredPane.sendMousePress(x = 100, y = 100)
-                        window.layeredPane.sendMouseRelease(x = 100, y = 100)
-                        awaitIdle()
-
-                        assertEquals(1, backgroundClickCount)
-                    }
-                } finally {
-                    window.dispose()
-                }
-            }
-        }
-    }
-
-    @Test
     fun unfocusablePopupLayer_withComponentLayerType_inComposePanel_isSizedCorrectly() {
         ComposeFeatureFlags.layerType.withOverride(LayerType.OnComponent) {
             ComposeFeatureFlags.useSwingGraphicsInComposePanel.withOverride(true) {
